Add 0.5U height support for rack items
Change item height_u from int to float to support half-U heights
common in 10" racks (patch panels, brush strips). Rack grid uses
half-U rows (each CSS grid row = 0.5U) for precise placement.

Backend: migration 006 changes column type, validation allows
multiples of 0.5, overlap detection uses float arithmetic.
Frontend: half-U grid rendering, step=0.5 on height input,
0.5U items span 1 grid row, 1U items span 2 grid rows.

+++ backend/app/models.py
@@ -236,7 +236,7 @@ class AppAuthStatus(BaseModel):
 
 class RackItemInput(BaseModel):
     position_u: int
-    height_u: int = 1
+    height_u: float = 1
     device_type: str = "other"
     label: str
     power_watts: float = 0.0
@@ -249,7 +249,7 @@ class RackItemInput(BaseModel):
 class RackItem(BaseModel):
     id: int
     position_u: int
-    height_u: int
+    height_u: float
     device_type: str
     label: str
     power_watts: float
@@ -278,7 +278,7 @@ class Rack(BaseModel):
     location: str
     items: list[RackItem] = []
     total_power: float = 0.0
-    used_u: int = 0
+    used_u: float = 0
 
 
 class RackSummary(BaseModel):
@@ -288,7 +288,7 @@ class RackSummary(BaseModel):
     height_u: int
     location: str
     item_count: int = 0
-    used_u: int = 0
+    used_u: float = 0
     total_power: float = 0.0
